begin factoring for parallel input history
